More tidying/deduping
